requirements.txt
paho-mqtt==2.1.0
tinytuya==1.15.1
pydantic==2.10.1
PyYAML==6.0.2
jinja2==3.1.4